fix: move steam resolution logic to shared module to fix bot import error
